# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BaZi v8.0",
    "version": "v1",
    "x-api-versions": {
      "v1": {
        "description": "初始版本（已废弃，将于 2026-12-31 停止支持）",
        "deprecated": true
      },
      "v2": {
        "description": "当前稳定版本：新增 output_format 精简响应、meta 信封（R38）、批量排盘（R39）",
        "deprecated": false
      }
    }
  },
  "paths": {
...
    "/api/v2/batch/verify": {
      "post": {
        "summary": "Api Batch Verify",
        "description": "N5.03 批量验证端点.\n\n- items 最多 50 条\n- ENGINE_V2=false 时返回 501（R40）\n- ThreadPoolExecutor(BATCH_WORKERS) 并行\n- 整体超时 60s，单条超时 5s\n- 部分成功返回（partial results on timeout）",
        "operationId": "api_batch_verify_api_v2_batch_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchVerifyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchVerifyRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/VerifyRequest"
            },
            "type": "array",
            "title": "Items",
            "description": "批量请求列表，最多 50 条"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "BatchVerifyRequest",
        "description": "N5.03 批量验证请求，最多 50 条"
      },
      "BatchVerifyResponse": {
        "properties": {
          "results": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Results",
            "description": "成功结果列表（有序，与 items 对应）"
          },
          "failed": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Failed",
            "description": "失败列表 [{index: int, error: str}]"
          }
        },
        "type": "object",
        "title": "BatchVerifyResponse",
        "description": "N5.03 批量验证响应"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VerifyRequest": {
        "properties": {
          "dt": {
            "type": "string",
            "format": "date-time",
            "title": "Dt",
            "description": "Input datetime; aware or naive ISO-8601"
          },
          "lon": {
            "type": "number",
            "title": "Lon",
            "description": "Longitude in degrees within supported range"
          },
          "mode": {
            "type": "string",
            "enum": [
              "dual",
              "single"
            ],
            "title": "Mode",
            "description": "Requested mode",
            "default": "dual"
          },
          "solar_time_enabled": {
            "type": "boolean",
            "title": "Solar Time Enabled",
            "description": "Enable solar-time adjustment when true",
            "default": false
          },
          "tz": {
            "type": "string",
            "title": "Tz",
            "description": "Timezone used only when dt is naive",
            "default": "Asia/Shanghai"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "male",
                  "female"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender",
            "description": "Gender for dayun direction: male/female"
          },
          "city_tier": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "一线",
                  "新一线",
                  "其余"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "City Tier",
            "description": "城市层级，影响财富数额地区系数：一线×1.8 / 新一线×1.2 / 其余×1.0 (M3.03)",
            "default": "其余"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "金融IT",
                  "教育公务",
                  "其余"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry",
            "description": "行业，影响财富行业系数：金融IT×1.5 / 教育公务×0.8 / 其余×1.0 (M3.03)",
            "default": "其余"
          }
        },
        "type": "object",
        "required": [
          "dt",
          "lon"
        ],
        "title": "VerifyRequest",
        "description": "验证请求"
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "标准错误代码",
                "example": "AUTH_001"
              },
              "message": {
                "type": "string",
                "description": "错误消息",
                "example": "Invalid credentials"
              },
              "details": {
                "type": "object",
                "description": "额外错误详情",
                "example": {}
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      }
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "AUTH_004",
                "message": "Missing authorization token"
              }
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "AUTHZ_001",
                "message": "Permission denied"
              }
            }
          }
        }
      },
      "NotFoundError": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "RESOURCE_001",
                "message": "Resource not found"
              }
            }
          }
        }
      },
      "ValidationError": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "VALIDATION_001",
                "message": "Validation failed",
                "details": {
                  "field": "email",
                  "reason": "Invalid format"
                }
              }
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "SYSTEM_001",
                "message": "Internal server error"
              }
            }
          }
        }
      }
    }
  }
}